Replaced request with Laravel Data & Added Policy

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,21 +3,15 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use BalajiDharma\LaravelAdminCore\Requests\StoreCategoryRequest;
-use BalajiDharma\LaravelAdminCore\Requests\UpdateCategoryRequest;
+use BalajiDharma\LaravelAdminCore\Actions\Category\CategoryCreateAction;
+use BalajiDharma\LaravelAdminCore\Actions\Category\CategoryUpdateAction;
+use BalajiDharma\LaravelAdminCore\Data\Category\CategoryCreateData;
+use BalajiDharma\LaravelAdminCore\Data\Category\CategoryUpdateData;
 use BalajiDharma\LaravelCategory\Models\Category;
 use BalajiDharma\LaravelCategory\Models\CategoryType;
 
 class CategoryController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware('can:category list', ['only' => ['index', 'show']]);
-        $this->middleware('can:category create', ['only' => ['create', 'store']]);
-        $this->middleware('can:category edit', ['only' => ['edit', 'update']]);
-        $this->middleware('can:category delete', ['only' => ['destroy']]);
-    }
-
     /**
      * Display a listing of the resource.
      *
@@ -25,6 +19,7 @@ class CategoryController extends Controller
      */
     public function index(CategoryType $type)
     {
+        $this->authorize('adminViewAny', Category::class);
         $items = (new Category)->toTree($type->id, true);
 
         return view('admin.category.item.index', compact('items', 'type'));
@@ -37,6 +32,7 @@ class CategoryController extends Controller
      */
     public function create(CategoryType $type)
     {
+        $this->authorize('adminCreate', Category::class);
         $item_options = Category::selectOptions($type->id, null, true);
 
         return view('admin.category.item.create', compact('type', 'item_options'));
@@ -47,13 +43,10 @@ class CategoryController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function store(StoreCategoryRequest $request, CategoryType $type)
+    public function store(CategoryCreateData $data, CategoryType $type, CategoryCreateAction $categoryCreateAction)
     {
-        if (! $request->has('enabled')) {
-            $request['enabled'] = false;
-        }
-
-        $type->categories()->create($request->all());
+        $this->authorize('adminCreate', Category::class);
+        $categoryCreateAction->handle($data, $type);
 
         return redirect()->route('admin.category.type.item.index', $type->id)
             ->with('message', 'Category created successfully.');
@@ -66,6 +59,7 @@ class CategoryController extends Controller
      */
     public function edit(CategoryType $type, Category $item)
     {
+        $this->authorize('adminUpdate', $item);
         $item_options = Category::selectOptions($type->id, $item->parent_id ?? $item->id);
 
         return view('admin.category.item.edit', compact('type', 'item', 'item_options'));
@@ -76,13 +70,10 @@ class CategoryController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(UpdateCategoryRequest $request, CategoryType $type, Category $item)
+    public function update(CategoryUpdateData $data, CategoryType $type, Category $item, CategoryUpdateAction $categoryUpdateAction)
     {
-        if (! $request->has('enabled')) {
-            $request['enabled'] = false;
-        }
-
-        $item->update($request->all());
+        $this->authorize('adminUpdate', $item);
+        $categoryUpdateAction->handle($data, $item);
 
         return redirect()->route('admin.category.type.item.index', $type->id)
             ->with('message', 'Category updated successfully.');
@@ -96,6 +87,7 @@ class CategoryController extends Controller
      */
     public function destroy(CategoryType $type, Category $item)
     {
+        $this->authorize('adminDelete', $item);
         $item->delete();
 
         return redirect()->route('admin.category.type.item.index', $type->id)
